feat: implement named NT sections

+++ crates/kernel/syscall/src/nt.rs
@@ -408,7 +408,7 @@ pub enum NtObjectCall {
     ResetEvent { handle: u32, previous: Option<UserPtr<i32>> },
     WaitEvent { handle: u32, alertable: u32, timeout: Option<UserPtr<i64>> },
     WaitMultiple { count: u32, handles: UserPtr<u32>, wait_type: u32, alertable: u32, timeout: Option<UserPtr<i64>> },
-    CreateSection { handle: UserPtr<u32>, desired_access: u32, size: u64, protect: u32, attributes: u32, file: u32 },
+    CreateSection { handle: UserPtr<u32>, desired_access: u32, size: u64, protect: u32, attributes: u64, file: u32 },
     MapViewOfSection { section: u32, process: u64, base: UserPtr<u64>, offset: u64, size: UserPtr<u64>, protect: u32 },
     UnmapViewOfSection { process: u64, base: u64 },
     UnmapViewOfSectionEx { process: u64, base: u64, flags: u32 },
@@ -1368,7 +1368,7 @@ pub fn decode_object(call: NtCall) -> Result<NtObjectCall, Errno> {
         }),
         NtService::CreateSection => Ok(NtObjectCall::CreateSection {
             handle: UserPtr::new(a.a0)?, desired_access: a.a1 as u32, size: a.a2,
-            protect: a.a3 as u32, attributes: a.a4 as u32, file: a.a5 as u32,
+            protect: a.a3 as u32, attributes: a.a4, file: a.a5 as u32,
         }),
         NtService::MapViewOfSection => Ok(NtObjectCall::MapViewOfSection {
             section: a.a0 as u32, process: a.a1, base: UserPtr::new(a.a2)?, offset: a.a3,
